refactor: delete dead code in src/tools/contacts/
Inline contact tool implementations (contact-search, contact-upsert,
contact-merge) directly into bundled-skills/contacts/tools/ and update
test imports. The old src/tools/contacts/ directory contained unregistered
handler stubs that only served as an unnecessary indirection layer.
